diskfilter: implementation of processing no metadata recorded in PV
If one PV underlying the root LV is created with no metadata, such as

pvcreate --metadatacopies 0 /dev/sda

then we could get a lot of error messages when generating a new
configuration file.

Generating grub configuration file ...
error: unknown LVM metadata header.
error: unknown LVM metadata header.
/usr/sbin/grub2-probe: warning: Couldn't find physical volume `pv1'.
Some modules may be missing from core image..
(For details, please refer to
 https://bugzilla.suse.com/show_bug.cgi?id=1027526)

When one labelled PV which dose not have any metadata is found, we put
it into a global grub_detached_pv list. and we search all the PVs in the
current array list to check if it is a member of grub_detached_pv list.
So we can know if the PV is really missing or just without metadata.

Signed-off-by: Lidong Zhong <lzhong@suse.com>